src/requirements.txt
python-telegram-bot
python-dotenv
psutil
speedtest-cli